Update charge_point.py
Adding MQTT_USERNAME and MQTT_PASSWORD.

I'm unsure if the if statement is needed as I'm unable to test the build. If passing empty username and password work when it's a non password protected MQTT, we could remove and keep only the portion with the username and password in the connection string.